pyproject.toml
[tool.poetry]
name = "crewai roadmap"
version = "0.1.0"
description = ""
authors = ["Biwas Bhandari <biwas2059@gmail.com>"]
readme = "README.md"
packages = [
    { include = "crews" }  # Include the crews directory
]

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai = "^0.76.9"
crewai-tools = "^0.13.4"
python-dotenv = "^1.0.1"
supabase = "^2.10.0"
uvicorn = "^0.32.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
